requirements.txt
Django==1.8.4
django-extensions==1.5.5
django-filter==0.11.0
django-rest-swagger==0.3.4
djangorestframework==3.2.2
PyYAML==3.11
six==1.9.0
